Add docs run options
